Changes HAL to return -1 to 1 for joysticks (#35)
First part of the HAL changes. Returns -1 to 1 for joysticks instead of
-128 to 127. Implemented by replacing the old structure with a new
structure that uses floats instead of shorts.

+++ hal/include/HAL/HAL.hpp
@@ -174,7 +174,7 @@ static const size_t kMaxJoystickPOVs = 12;
 
 struct HALJoystickAxes {
   uint16_t count;
-  int16_t axes[kMaxJoystickAxes];
+  float axes[kMaxJoystickAxes];
 };
 
 struct HALJoystickPOVs {
